+++ lib/infopark_kickstarter/rake/integration_task.rb
@@ -73,6 +73,8 @@ module InfoparkKickstarter
           'cms:widget:slider --example',
           'cms:widget:login --example',
           'cms:widget:slideshare --example',
+          'cms:widget:column --columns=2 --example',
+          'cms:widget:column --columns=3 --example',
         ]
 
         generators.each do |generator|