+++ tmp/test_app/app/models/column3_widget.rb
@@ -1,0 +1,10 @@
+class Column3Widget < Obj
+  cms_attribute :column_1, type: :widget
+  cms_attribute :column_2, type: :widget
+  cms_attribute :column_3, type: :widget
+
+  # Most CMS objects are either a page or a box. In order for them to
+  # have common behavior, uncomment one of the following lines.
+  # include Page
+  include Widget
+end